requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.0
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
gunicorn==20.1.0
idna==3.4
PyYAML==6.0
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.26.14